Fixes tracking indirect dependencies & __init__.py
Also adds test coverage.

We use a _currently_importing_package to keep track of which package is
currently performing an import, so that indirect imports are still
associated with that package (for dependency-tracking purposes) although
they are not originating from it.

This field is also set in Package#load() so that the PackageManager can
find this package, that is not yet stored in the list of packages.