Add tests for translations between db versions.
To make things easier, we embed a deep_eq_test method in the db domain objects.  This test recurses on lists and referenced objects so that it checks entire objects.  Whether this should be a permanent piece of the domain object or only enabled for testing should be considered, but we are including it in the standard template for now.

Because certain fields or child objects may change or be added in later versions, we have an alternate_tests dictionary that allows a user to specify a different test than the standard assertEqual; None indicates that the test should be skipped.  In addition, vistrails.db.services.io.TestTranslations has a get_alternate_tests method which builds the dictionary based on the versions where a workaround is required.  That way, more recent versions can be more rigorous.

TODO: Improve coverage.  Either use more data or manufacture some data to ensure that all fields are covered.

vistrails/db/bin/templates/domain.py.mako:
  - add deep_eq_test to facilitate easier equality testing

vistrails/db/services/io.py:
  - TestTranslations: new class to test translations between versions
  - TestTranslations.get_alternate_tests: builds alternate tests based on versions where they are needed
  - TestTranslations.run_vistrail_translation_test: new method
  - TestTranslations.run_workflow_translation_test: new method
  - TestTranslations.run_log_translation_test: new method
  - TestTranslations.run_registry_translation_test: new method
  - TestTranslations: various tests for specific versions

vistrails/db/versions/__init__.py:
  - version_map and rev_version_map made module-level to allow more general use
  - get_version_path: new method
  - translate_object: move version_map and rev_version_map outside method to allow more general use, remove unused path list

vistrails/db/versions/v1_0_4/domain/auto_gen.py:
  - regenerated using the new template with the deep_eq_test method