<feature>Start migrating parameter explorations through upgrades</feature> <ticket>#695</ticket>
This is still WIP as the parameter explorations are migrated, but getting to them is still very clunky with upgradeDelay=True.  We only migrate the exploration when we get a totally 1-1 remap between module ids.  In upgradeDelay mode (default), the new version is not available in the version tree until the user changes the pipeline so to get to the upgraded parameter exploration(s), a user must make an arbitrary change, go to the version tree, and select the upgraded version (with migrateTags=True and a tagged version, this is a bit easier).  Then, when switching into parameter exploration mode, they should see the upgraded parameter explorations.

The upgrade is implemented by using do_copy with the generated module remap obtained from paired add/delete module operations.

vistrails/core/vistrail/controller.py:
  - VistrailController.__init__: init _delayed_paramexps
  - VistrailController.flush_delayed_actions: also flush delayed parameter explorations
  - VistrailController.get_upgrade_module_remap: new method
  - VistrailController.handle_invalid_pipeline: add logic for upgrading parameter explorations