@parallelizable doesn't enable thread by default
This marker will now be used to indicate that a module can be run in the
interpreter thread, which is separate from Qt's UI thread.

Modules that should be run on the UI thread should either not use it, or
for more complex modules, use it and explicitely use
self.moduleInfo['unparallelizable_execution'](...) to call the unsafe
function on the UI thread.

Note that modules that interact with the task system MUST use
@parallelizable, else their do_compute() method is automatically called
on the UI thread where the task system should NOT be used.