MplFigure#updateUpstream() sets a figure on plots
MplPlot now has the figure to be used as an attribute. This allows the
modules to not be dependent on the order of execution (i.e. having plots
executed after the figure's updateUpstream() and its compute(), without
any other figure doing anything), which the work on parallelization
requires, and to detect the case when a single plot is connected to
multiple figures (which would result in the second figure to execute
being empty).