<bugfix>Fixed tricky upgrade bug</bugfix>
When we search for a version that is specified in an annotation, we need to be careful about types.  The recursive get_upgrade fails because the recursed value was a string not a long.  Hence, the dict lookup failed and we didn't think we had a secondary upgrade even though one existed.

vistrails/core/vistrail/vistrail.py:
  - Vistrail.get_upgrade: recurse on value converted to long