WIP: Improve inline module settings and port specification
Continuing with the idea from the previous thread, we implement
namedtuples for ModuleSettings, Ports, etc.  The major benefit here is
a more readable inline shorthand for settings and ports.  Specifying
ports and settings inline is preferred to using initialize because
everything about a module is stored in a single location.  Before now,
the drawback was that the syntax was a collection of tuples with
dictionaries.  The namedtuples make it possible to use kwargs to make
it much clearer what is going on.

Here is an example of the syntax from the File module:

class File(Path):
    _settings = ModuleSettings(constantSignatureCallable=path_parameter_hasher,
                               constantWidget=("%s:FileChooserWidget" % \
                                               constant_config_path))
    _input_ports = [Port("value", "File"),
                    Port("create_file", "Boolean", optional=True)]
    _output_ports = [Port("value", "File"),
                     Port("self", "File", optional=True),
                     Port("local_filename", "String", optional=True)]

    ...

Note that it may make sense to differentiate between input and output
ports as defaults, entry_types, etc. do not make sense for output
ports.  In addition, Port is now assumed to be a simple (non-compound)
port since these are most likely to be the norm for most developers.
Then, "default", "entry_type", and "label" are all singular and
"values" is only a single list of values (not a list of lists).
CompoundPort is used for compound ports like (Integer, Float)-type
ports, and can be specified in the old syntax or using PortItems.

The one drawback I have found is that "self"-style ports cannot be
specified using classes.  For example, in the above example,
Port("value", "File") cannot be written as Port("value", File) since
this is in the definition of the File class.  However, Port("value",
String) would be allowed.  Personally, I think the string-based
version is much better anyway, especially for cross-package
references.

This is still WIP and will likely change.  It would be nice to unify
the actual registry calls with the config objects so that changing one
automatically updates the other.  Also, it would be nice to improve
error messages.

vistrails/core/modules/basic_modules.py:
  - move everything in initialize to _settings, _input_ports, and
    _output_ports

vistrails/core/modules/config.py:
  - ModuleSettings: new namedtuple
  - Port: new namedtuple
  - OutputPort: new namedtuple
  - CompoundPort: new namedtuple
  - PortItem: new namedtuple
  - OutputPortItem, new namedtuple

vistrails/core/modules/module_registry.py:
  - ModuleRegistry.auto_add_ports: add translations for
    Port/CompoundPort entries in _input_ports/_output_ports
  - ModuleRegistry.auto_add_module: add support for ModuleSettings in
    Module._settings
  - ModuleRegistry.add_module: don't use fetch for defaults to allow
    ModuleSettings defaults' to work, just use None and add defaults
    after fetch

vistrails/core/modules/sub_module.py:
  - move everything in initialize to _settings, _input_ports, and
    _output_ports

vistrails/core/modules/vistrails_module.py:
  - move everything in initialize (of basic_modules.py) to _settings,
    _input_ports, and _output_ports