WIP: Allow default/values to be set by python value
One annoyance of specifying a default value or a list of enumerated
values for a port is that they must be specified as strings.  This is
because these must be serializable values.  However, this may be
inconvenient and not easy to remember for very simple types like
bools, floats, etc.  This commit introduces a new method
convert_port_val that attempts to convert a value (from the default or
values list) to the necessary string form from a possible python
value.  If the string value is specified, it also attempts to check
whether this is the serialized or python form (should hopefully be the
same in this case!).  Since Constants are required to have
translate_to_string and translate_to_python methods, this is possible
for any Constant which is also required to use default/values.

This is very WIP as the code only works for Port()-style definitions
right now.

vistrails/core/modules/basic_modules.py:
  - FileSink.overwrite: set default as python value True instead of str

vistrails/core/modules/module_registry.py:
  - ModuleRegistry.convert_port_val: new method
  - ModuleRegistry.auto_add_ports: add some scaffolding to test
    convert_port_val, needs to be refactored to be more generally
    applicable