Persistence sets up a Git username if none is set
If user.name or user.email are not set, as might be the case if the user
didn't use Git outside of VisTrails, committing from VisTrails will
fail.

This changeset sets up the required variables in the persistent_files
directory's local configuration so that the commit works.

core.system.current_user() is used to get the user name.