<bugfix>Fix issue with subworkflows using disabled packages</bugfix> <ticket>#728</ticket>
The issue was that subworkflow loading only stepped through one level
of invalid pipeline handling.  Thus, the first level catches and deals
with the MissingPackage exception, but trying to use the subworkflow
after that fails because it was not upgraded to match the new version
of the package.  This code adds a second level to catch a second
InvalidPipeline exception.  Also, force_no_delay must be True for any
upgrades to be correctly used.

The commit also changes to use an independent controller instead of
attempting to use the current controller by switching the current
vistrail and id_scope.

vistrails/core/vistrail/controller.py:
  - VistrailController.add_abstraction_to_registry: add second level
    of error handling and use new controller