Restructure developer documentation.
Change the structure of the developer documentation so it is easier for developers to locate information; there is now a section for Package Specification, Module Specification, and Port Specification.  Specialized topics are moved to the bottom for now, but need to be better organized.  Also, I broke the flow since it used to Afront example, but this needs to be updated anyway.  Also add API Reference to TOC.