Add script to generate VisTrails package documentation for sphinx.
The script generates a rst file that contains all the modules for each
package as well as the input and output ports (along with their
signatures) and any developer-specified documentation for the package,
modules, and ports.  Ideally, for each module, we would also include
links to example vistrails that contain that module type.

scripts/generate_pkg_doc.py:
  - new script to generate documentation that can be processed by sphinx