Fix issue with defaulting the package identifier.
The registry's self._current_package can be None if no package is
being loaded.  In such cases, we need to detect this and flag the
error if the user has not specified the package identifier as an
argument to add_module.  This should normally only be an issue for
tests.

vistrails/core/modules/module_registry.py:
  - ModuleRegistry.add_module: fix issue when defaulting package
    identifier and version