Change naming across application to match new configuration.
*** NOTE *** This supposes an updated ~/.vistrails/startup.xml file so I
wouldn't recommend running it with your normal startup.xml as it can
rewrite that file. Translations should fix this issue later.

This is mostly name switches, although also moves logs to a "logs"
subdirectory in ~/.vistrails since the logFile config that converted
to a directory and then named it according to version was obfuscated.
Also uses the new command-line parser in configuration.

The test suite passes modulo the TestImports test which has an issue
with the Mac binary.