Adds ModuleWasSuspended
Previously, if a module had two upstream connections to the same module
and it suspended, the second update() call from updateUpstream() would
raise ModuleHadError, which wasn't caught, and masked the first
ModuleSuspended (because we didn't get a chance to re-raise).

The upstream will now raise ModuleWasSuspended on the second call, which
will be caught, allowing the ModuleSuspended to be re-raised.
ModuleWasSuspended is only raised if there is no ModuleSuspended.