Fix schema issue where loop_exec still had an inverse reference to itself.
To support the way the SQL schema is written, we maintain inverse
references, but we also use the forward references and both must
exist.  With the change to add loop_iteration, we do not have the
loop_exec -> loop_exec structure any longer.  This means that the
loop_exec property in the "parent" choice should also be removed.  The
code was looking for the forward reference to loop_exec so it could do
the right append when reading in from a relational database, and that
didn't exist.

Also, this deletion of a previous valid nesting may break any
vistrails written with the v1.0.3 schema so it may be best to put this
in the v1.0.4 schema.  (We can still release v2.1 with a v1.0.4 schema
if necessary, but we will need translation code for the log.)