Overrides output ports for Map
module_to_serialized_pipeline() gets the connected output ports. This is
what will be sent over the wire along with the module, and that
execute_serialized_pipeline() will use to build the outputs dictionary.

However when using Map, the 'self' output is connected although it causes
problems (Modules are not pickle-safe!), and the output ports that are
actually needed (the OutputPort input of Map) is not set (because not
connected). This adds a serialized_outputports attribute on Module that
can be used to override what outputs are actually needed.

Ticket: 797