Initial	version	of a recursive remap_module and new specification.
Unforseen issues include reliance of create_function and
create_new_connection on the registry (and thus on the module
descriptor we are upgrading to actually existing which isn't the case
for an intermediate upgrade).  Got around this by creating dummy ports
and port specs and allowing the controller to deal with these.  Also
needed to duplicate the pipeline and apply the intermediate upgrades
because subsequent calls reference the remapped modules.  Finally, the
commit adds a test package and a test of the upgrade.