Prevents enabling package from making another fail
When enabling a package, another one gets selected in the 'available
packages' list which triggers its loading. If __init__.py fails or takes
a long time (for example: the package doesn't use a separate init.py),
this is an issue.

This patch makes it so no other package in the 'available' list gets
selected.

Fixes: 783