Fixes handling jobs in loops
Catching ModuleSuspended in weird places to set loop_iteration seemed
error-prone and wrong. Instead, we get the iteration from the
LogController when we handle it in the interpreter code.

The major issue here is that this was not compatible with the
multithreaded interpreter, since in this case exceptions do not
propagate to downstream modules (they reach the interpreter immediately).