Reworks JobMixin to use self.signature
The subpipeline signature now acts as the job identifier. This means
that we no longer need to run the upstream modules to get the parameter
values so we can compute the identifier, so we don't. The upstream is
only computed when a new job needs to be submitted (this is similar to
what persistence does).

When a FoldWith module (such as Map) is used, the upstream modules will
still be executed, because they are not actually upstream of JobMixin.
This is unfortunate.