Do not ignore pending module moves when saving
Module moves are kept 'pending' until some other action occur, at which
point they are recorded as a single action. However, if no action occur
before saving (or exiting) these are not written.

This makes a module move mark the controller as modified (so it appears
with a '*', and VisTrails prompts to save on exit) and makes saving
flush the pending moves first.