Change QPipelineScene.pipeline to current_pipeline
Only QPipelineScene.setupScene used the pipeline field, and the
controller was linked to the current_pipeline field which is stored on
QPipelineScene, too.  This difference caused issues with code that
used self.pipeline and self.current_pipeline when the two were out of
sync (e.g. the grouping code).  Now, we just use current_pipeline
instead of pipeline.

Ticket: 869