Makes do_version_switch() to current validate
Previously, no validation would occur if we were changing to the already
selected version. This was an important optimization. However, this
meant that calling perform_action() then change_selected_version(),
neither of them would validate.

This makes change_selected_version() to current pipeline (re)validate if
the current pipeline was invalid (or not yet validated). This means that
changing to the current version is still fast, but errors will be raised
if present.

The behavior that changes is that you will get exceptions when changing
to an invalid pipeline if it was previously selected.