Different approach -- two loggers
The logger that the 'debug' module logs on is now vistrails.logger. It
is internal and API users should NOT configure it (change level or
attach filters/handlers to it). It is controlled by the 'debug' module.

This internal logger re-logs to a 'vistrails' logger, which is meant to
be configured by API users if needed -- changing the level of this one
won't break anything.