Updated source build script to be runnable by a buildbot
Removed archiving, git checkout/exporting, and hash/changelog generation.

The script now uses the existing git tree. You need to use 'git clean' and
run dist/common/prepare_release.py to get a clean and up-to-date tar file.

It will upload to sourceforge as nightly if username and keydir is provided.

It can do a release build by copying the resulting tar file, or by
modifying the script to upload to a new binary location.