Don't use get_vistrails_default_pkg_prefix()
The whole point of identifiers is to be a unique immutable string
identifying a module. Dynamically computing it defeats the purpose.

They should not change and code should rely on that fact. VisTrails has
already been made tolerant to changes in identifiers via the
old_identifiers and can_handle_identifier() mechanisms.