Adds ipython_mode(True|False) to the API
This allows the user to explicitely selects if he wants to use the
notebook. It is similar to matplotlib's `%matplotlib inline` magic.

Note that the current default is to still enable IPythonMode if we think
we are running from the notebook, which right now only means that the
__main__ module exists but has no __file__ attribute, and that IPython
is installed.