wrapper/specs.py: Generalized attributes
All attributes are now handled using the "args" specification.

The patching step can now work on this specification and
it will automatically work for subclasses.

TODO:
How to handle list types in subclasses? We should have a dict type
that can patch each item in the dict. AlternatePortSpec will need this.