UpgradeWorkflowHandler: Assume first upgraded module is the upgraded one
A module can be upgraded to multiple modules. The new upgrade
code allows intermediate upgrades and needs to select
the correct module to upgrade to the final version.

Previously the code selected the last added module.
This breaks some modules that upgrades to multiple modules.
We now select the first added module instead.
This seems to work, but could still be brittle.