Allow new_module ModuleDescriptor upgrade to refer to an outdated module
Before we always tried to build a new valid ModuleDescriptor for
new module upgrades. Now we use the given descriptor so that it does not
need to refer to an up-to-date module type.

Also, now ModuleDescriptor.package_version is respected so that
we do not replace the descriptor package_verasion with the current
package_version.