Added operations
An operation module outputs a method signature and arguments that is
used by the class module to apply the method to the class instance.

This can be used to create better data-flow. Mutating operations
can now be applied in a single step to the class instance.

Additionally, matplotlib and bokeh can use this to construct
operations that connect to a canvas, rather than the other
way around, which is the preferable direction because the
canvas connects to the output cell.