Adds union port
A union port is a list of ports sharing the same sort_key.
They are shown as a single port in the ports pane.
The name of the union is the common name prefix of the union ports.
Adding/showing a union port will first prompt which type to use.