Adds portspec.union and new schema version 1.0.5
portspec.union (string) is now used to identify unions
in place of sort_key. The value of .union is used instead
of name in the gui as the visual identifier. portspec.name is
not visible in the pipeline gui, but may still be used in
parts of the interface that does not yet understand unions.