Make sure field order is maintained from spec
This is important because if we copy an entity, the code needs to be
aware of any changes to ids that may be in the same object
(e.g. DBWorfklowExec.item_execs references ids in
DBWorkflowExec.machines via the machine_id foreign key, but before the
change, machines were added after item_execs). This is only an issue
when choices need to be after properties.